requirements.txt
streamlit
PyPDF2
scikit-learn
numpy
pandas
google-generativeai
concurrent-futures